requirements.txt
psycopg2-binary
clickhouse-connect
python-dotenv
pandas

